Use package resources in dataset accessors
Currently, the data sets contain references to specific files, relative to an arbitrary location, which won't work when distributed as a Python package. The correct way to reference data included in a Python packages is to use importlib.resources (Python 3.7) or its backport.

This adds a new abstraction that allows some of the dataset-related functions to be passed package resources (or to at least use them as default values).

PiperOrigin-RevId: 319982283