Move tests into a tests/ folder
This will make packaging the OSS release with PyPA best practices easier, but it is not *required* to package this correctly. It is generally preferable to distribute the tests but not install them, which is considerably harder to achieve when the tests live next to the code under test instead of in a separate tests/ folder.

This also migrates `tests/file_utils_test.py` to using `pytest` as the test runner, because the test runner that blaze uses does not support relative imports, which is necessary to avoid requiring that the tests are run from the repo root.

PiperOrigin-RevId: 320197579